refactor(backend-service-game): refactor converters to be builders
