+++ packages/api/libraries/api-models/src/models/ApiSchemaId.ts
@@ -10,7 +10,6 @@ export enum ApiSchemaId {
   UserV1 = 'https://onegame.schemas/api/v1/users/user.json',
   UserIdUpdateQueryV1 = 'https://onegame.schemas/api/v1/users/user-id-update-query.json',
   UserCreateQueryV1 = 'https://onegame.schemas/api/v1/users/user-create-query.json',
-  ErrorV1 = 'https://onegame.schemas/api/v1/errors/error.json',
   NonStartedGameV1 = 'https://onegame.schemas/api/v1/games/non-started-game.json',
   NonStartedGameSlotV1 = 'https://onegame.schemas/api/v1/games/non-started-game-slot.json',
   GameV1 = 'https://onegame.schemas/api/v1/games/game.json',
@@ -27,6 +26,7 @@ export enum ApiSchemaId {
   ActiveGameV1 = 'https://onegame.schemas/api/v1/games/active-game.json',
   ActiveGameSlotV1 = 'https://onegame.schemas/api/v1/games/active-game-slot.json',
   ActiveGameSlotCardsV1 = 'https://onegame.schemas/api/v1/games/active-game-slot-cards.json',
+  ErrorV1 = 'https://onegame.schemas/api/v1/errors/error.json',
   WildDraw4CardV1 = 'https://onegame.schemas/api/v1/cards/wild-draw-4-card.json',
   WildCardV1 = 'https://onegame.schemas/api/v1/cards/wild-card.json',
   SkipCardV1 = 'https://onegame.schemas/api/v1/cards/skip-card.json',
