+++ packages/backend/services/backend-service-game/src/cards/application/builders/CardFromCardV1Builder.ts
